+++ __snapshots__/flex-job.test.js
@@ -80,7 +80,7 @@ exports['flex job by specifying credentials as options when invalid and existent
 [debug] Logging in user: johnDoe@mail.com
 [debug] Request:  POST http://localhost:3234/session
 [debug] Response: POST http://localhost:3234/session 401
-[error] InvalidCredentials: Credentials are invalid. Please authenticate.
+[error] InvalidCredentials: Invalid e-mail and/or password.
 
 `