+++ __snapshots__/flex-recycle.test.js
@@ -129,7 +129,7 @@ exports['flex recycle by specifying credentials as options when invalid and exis
 [debug] Logging in user: johnDoe@mail.com
 [debug] Request:  POST http://localhost:3234/session
 [debug] Response: POST http://localhost:3234/session 401
-[error] InvalidCredentials: Credentials are invalid. Please authenticate.
+[error] InvalidCredentials: Invalid e-mail and/or password.
 
 `