+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+* Suggest command when token is expired
+* Use backend error when HTTP status code is 401
+* Apply Kinvey lint rules and fix lint errors
+
 ## 4.4.0-alpha.0 (2018-12-14)
 * Add initial websites support